shift configuration into an isolated directory and split out the various configuration interfaces into multiple files
