+++ src/configuration/snapshot-configuration.ts
@@ -1,0 +1,4 @@
+export interface SnapshotConfiguration {
+  widths?: [number],
+  'min-height'?: number,
+}
