+++ src/configuration/static-snapshots-configuration.ts
@@ -1,0 +1,5 @@
+export interface StaticSnapshotsConfiguration {
+  'base-url'?: string,
+  'snapshot-files'?: string,
+  'ignore-files'?: string,
+}
