+++ src/services/agent-service.ts
@@ -2,8 +2,9 @@ import * as bodyParser from 'body-parser'
 import * as cors from 'cors'
 import * as express from 'express'
 import {Server} from 'http'
-import { SnapshotOptions } from '../percy-agent-client/snapshot-options'
-import configuration, {SnapshotConfiguration} from '../utils/configuration'
+import configuration from '../configuration/configuration'
+import {SnapshotConfiguration} from '../configuration/snapshot-configuration'
+import {SnapshotOptions} from '../percy-agent-client/snapshot-options'
 import logger, {profile} from '../utils/logger'
 import {AgentOptions} from './agent-options'
 import BuildService from './build-service'
